+++ packages/clients/diplan/src/store/geoEditing/duplicatePolygons.ts
@@ -2,7 +2,11 @@ import { PolarActionContext } from '@polar/lib-custom-types'
 import { DiplanGetters, DiplanState } from '../../types'
 
 export const duplicatePolygons = ({
+  commit,
   dispatch,
 }: PolarActionContext<DiplanState, DiplanGetters>) => {
   dispatch('plugin/draw/setMode', 'none', { root: true })
+  commit('setDrawMode', 'duplicate')
+
+  // TODO on end: commit('setDrawMode', null)
 }