+++ packages/plugins/Draw/src/index.ts
@@ -6,6 +6,7 @@ import locales from './locales'
 import { makeStoreModule } from './store'
 
 export { getSnaps } from './store/createInteractions/getSnaps'
+export type { Mode } from './types'
 
 // NOTE: Currently no options are specified here, variable is kept for integrity until options are needed
 export default (options: DrawConfiguration) => (instance: Vue) =>